+++ src/box-client.ts
@@ -30,7 +30,12 @@ import Trash from './managers/trash';
 import Users from './managers/users';
 import WebLinks from './managers/web-links';
 import Webhooks from './managers/webhooks';
-import FileRequestsManager from './managers/file-requests-manager';
+import FileRequestsManager from "./managers/file-requests-manager";
+import ShieldInformationBarriers from "./managers/shield-information-barriers.generated";
+import ShieldInformationBarrierSegments from "./managers/shield-information-barrier-segments.generated";
+import ShieldInformationBarrierSegmentMembers from "./managers/shield-information-barrier-segment-members.generated";
+import ShieldInformationBarrierSegmentRestrictions from "./managers/shield-information-barrier-segment-restrictions.generated";
+import ShieldInformationBarrierReports from "./managers/shield-information-barrier-reports.generated";
 
 // ------------------------------------------------------------------------------
 // Typedefs and Callbacks
@@ -194,6 +199,11 @@ class BoxClient {
 	termsOfService: any;
 	storagePolicies: any;
 	signRequests: SignRequests;
+	shieldInformationBarriers: ShieldInformationBarriers;
+	shieldInformationBarrierSegments: ShieldInformationBarrierSegments;
+	shieldInformationBarrierSegmentMembers: ShieldInformationBarrierSegmentMembers;
+	shieldInformationBarrierSegmentRestrictions: ShieldInformationBarrierSegmentRestrictions;
+	shieldInformationBarrierReports: ShieldInformationBarrierReports;
 
 	/* prototype properties assigned below the class declaration */
 	collaborationRoles!: Record<string, CollaborationRole>;
@@ -262,6 +272,13 @@ class BoxClient {
 		this.termsOfService = new TermsOfService(this);
 		this.storagePolicies = new StoragePolicies(this);
 		this.signRequests = new SignRequests(this);
+		this.shieldInformationBarriers = new ShieldInformationBarriers(this);
+		this.shieldInformationBarrierSegments = new ShieldInformationBarrierSegments(this);
+		this.shieldInformationBarrierSegmentMembers = new ShieldInformationBarrierSegmentMembers(this);
+		this.shieldInformationBarrierSegmentRestrictions = new ShieldInformationBarrierSegmentRestrictions(this);
+		this.shieldInformationBarrierReports = new ShieldInformationBarrierReports(this);
+		// Legacy insensitive language
+		this.collaborationWhitelist = this.collaborationAllowlist;
 	}
 
 	/**